Renamed project from 'flutter_geolocator' to 'geolocator'
