--- ios/Classes/FlutterGeolocatorPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterGeolocatorPlugin : NSObject<FlutterPlugin>
-@end
