+++ ios/Classes/GeolocatorPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface GeolocatorPlugin : NSObject<FlutterPlugin>
+@end
