+++ ios/Classes/GeolocatorPlugin.m
@@ -1,0 +1,8 @@
+#import "GeolocatorPlugin.h"
+#import <geolocator/geolocator-Swift.h>
+
+@implementation GeolocatorPlugin
++ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
+  [SwiftGeolocatorPlugin registerWithRegistrar:registrar];
+}
+@end
