+++ ios/geolocator.podspec
@@ -2,15 +2,15 @@
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html
 #
 Pod::Spec.new do |s|
-  s.name             = 'flutter_geolocator'
+  s.name             = 'geolocator'
   s.version          = '0.0.1'
-  s.summary          = 'Geolocation plugin for Flutter (based on the excellent Geolocator plugin for Xamarin by James Montemagno)'
+  s.summary          = 'Geolocation plugin for Flutter'
   s.description      = <<-DESC
-Geolocation plugin for Flutter (based on the excellent Geolocator plugin for Xamarin by James Montemagno)
+Geolocation plugin for Flutter
                        DESC
-  s.homepage         = 'http://example.com'
+  s.homepage         = 'http://github.com/BaseflowIT/flutter-geolocator'
   s.license          = { :file => '../LICENSE' }
-  s.author           = { 'Your Company' => 'email@example.com' }
+  s.author           = { 'Baseflow' => 'hello@baseflow.com' }
   s.source           = { :path => '.' }
   s.source_files = 'Classes/**/*'
   s.public_header_files = 'Classes/**/*.h'
